+++ openfold3/entry_points/validator.py
@@ -158,7 +158,7 @@ class OutputWritingSettings(BaseModel):
 
     structure_format: Literal["pdb", "cif", "cif.gz"] = "cif"
     full_confidence_output_format: Literal["json", "npz"] = "json"
-    full_confidence_output_dtype: Literal["float16", "float32"] = "float32"
+    full_confidence_output_dtype: Literal["float16", "float32"] = "float16"
     write_features: bool = False
     write_latent_outputs: bool = False
     write_full_confidence_scores: bool = True
